Feat/paciente

+++ src/main/java/BackendC3/ClinicaOdontologica/repository/ITurnoRepository.java
@@ -3,5 +3,5 @@ package BackendC3.ClinicaOdontologica.repository;
 import BackendC3.ClinicaOdontologica.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITurnoRepository extends JpaRepository<Turno, Integer> {
+public interface ITurnoRepository extends JpaRepository<Turno, Long> {
 }
